examples/weather/flood_modeling/hydrographnet/requirements.txt
hydra-core
jaxtyping
matplotlib
mlflow>=2.1.1
networkx
nvtx
Pillow
psutil
requests
scipy
termcolor>=2.1.1
torch_geometric>=2.6.1
torch_scatter>=2.1.2
tqdm
wandb
